feat: adding sort capability to history table
