+++ packages/client/package.json
@@ -25,9 +25,9 @@
 	"dependencies": {
 		"@auth0/auth0-react": "2.2.4",
 		"@floating-ui/react": "0.26.12",
-		"@versini/ui-components": "5.14.5",
+		"@versini/ui-components": "5.16.1",
 		"@versini/ui-form": "1.3.0",
-		"@versini/ui-icons": "1.5.0",
+		"@versini/ui-icons": "1.6.0",
 		"@versini/ui-system": "1.3.0",
 		"autoprefixer": "10.4.19",
 		"postcss": "8.4.38",
